requirements.txt
atomicwrites==1.4.1
attrs==21.4.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.6.15
charset-normalizer==2.1.0
cloudscraper==1.2.60
colorama==0.4.5
greenlet==1.1.2
idna==3.3
importlib-metadata==4.12.0
iniconfig==1.1.1
Js2Py==0.71
packaging==21.3
Pillow==9.2.0
playwright==1.24.0
pluggy==1.0.0
py==1.11.0
pyee==8.1.0
pyjsparser==2.7.1
pyparsing==3.0.9
pytest==7.1.2
pytest-base-url==2.0.0
pytest-playwright==0.3.0
python-slugify==6.1.2
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
requests-toolbelt==0.9.1
six==1.16.0
soupsieve==2.3.2.post1
text-unidecode==1.3
tomli==2.0.1
typing_extensions==4.3.0
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.10
websockets==10.1
zipp==3.8.1
